--- src/features/creating-question/manual/customTextareaField.tsx
@@ -1,36 +1,0 @@
-import { FormControl, FormField, FormItem, FormLabel, FormMessage } from "@ui/form";
-import { Textarea } from "@ui/textarea";
-import { formLabelClasses, textAreaBold } from "./styles";
-import type { TCustomTextAreaField } from "./types";
-
-export const CustomTextAreaField = ({
-  control,
-  disabled,
-  name,
-  formLabel,
-  placeholder,
-}: TCustomTextAreaField) => {
-  return (
-    <div>
-      <FormField
-        control={control}
-        name={name}
-        render={({ field }) => (
-          <FormItem className="pt-8">
-            <FormLabel className={formLabelClasses}>{formLabel}</FormLabel>
-            <FormControl>
-              <Textarea
-                placeholder={placeholder}
-                disabled={disabled}
-                className={textAreaBold}
-                style={{ fontSize: "1.25rem" }}
-                {...field}
-              />
-            </FormControl>
-            <FormMessage />
-          </FormItem>
-        )}
-      />
-    </div>
-  );
-};
