--- src/features/creating-question/manual/index.ts
@@ -1,1 +1,0 @@
-export { FormForCreatingQuestionsYourself } from "./form-for-creating-questions-yourself";
