--- src/features/creating-question/manual/styles.ts
@@ -1,10 +1,0 @@
-export const textAreaBase = "text-center placeholder:text-sm";
-
-export const textAreaBold =
-  "text-center placeholder:text-sm font-extrabold placeholder:font-normal";
-
-export const formLabelClasses = "text-lg text-yellow-50";
-
-export const formActionContainerStyle = "mb-8 flex w-full justify-evenly pt-14";
-
-export const answersFieldContainerStyles = "mx-auto w-[85%]";
