+++ src/features/manual-creating-question/index.ts
@@ -1,0 +1,1 @@
+export { FormForCreatingQuestionsYourself } from "./ui/form-creating-questions-yourself";
