+++ src/pages/creating/manual/manual-question-creation.tsx
@@ -1,7 +1,7 @@
 import { formContainerStyles, headerStyles } from "./styles";
 import { TManualCreatingPage } from "./types";
 import { useTranslation } from "react-i18next";
-import { FormForCreatingQuestionsYourself } from "@features/creating-question/manual";
+import { FormForCreatingQuestionsYourself } from "@features/manual-creating-question";
 
 export const ManualCreatingPage = ({ courseName, themeName, themeID }: TManualCreatingPage) => {
   const { t } = useTranslation();
